requirements.txt
httpx>=0.27
beautifulsoup4>=4.12
lxml>=5.0
sqlalchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9
python-dotenv>=1.0
fake-useragent==2.2.0